Change logic of assassins

+++ Valtech_Task2(Ankh-Morpork game)/Data/Models/Assassins.cs
@@ -5,8 +5,8 @@ namespace Valtech_Task2_Ankh_Morpork_game_.Guilds.GuildMembers
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int MaxRange { get; set; }
-        public int MinRange { get; set; }
+        public decimal MaxRange { get; set; }
+        public decimal MinRange { get; set; }
         public bool IsOccupied { get; set; }
         public override string ToString() { return Name; }
     }
